refactor: break out items
